fix: gate browser manual-login setup consistently
